refactor: move ExportButton and OfflineBanner to chat folder
Components are used in ChatContainer, so they belong in components/chat/
