+++ components/chat/OfflineBanner.tsx
